flags extractions to avoid cycles
