+++ src/snowflake/cli/_plugins/stage/commands.py
@@ -30,7 +30,7 @@ from snowflake.cli._plugins.stage.diff import (
     compute_stage_diff,
     print_diff_to_console,
 )
-from snowflake.cli._plugins.stage.manager import OnErrorType, StageManager
+from snowflake.cli._plugins.stage.manager import StageManager
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.commands.flags import (
     ExecuteVariablesOption,
@@ -40,6 +40,7 @@ from snowflake.cli.api.commands.flags import (
     like_option,
 )
 from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
+from snowflake.cli.api.commands.structs import OnErrorType
 from snowflake.cli.api.console import cli_console
 from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.identifiers import FQN
